02-keras.md: adjust sections
All the items of the Deep Learning workflow are now second-level sections. The headings are now all lower-case for consistency.

The title of Section 4 reflects the order its main topics are introduced in (first build an architecture from scratch, then choose a pretrained model). The subsection 'Keras for Neural Networks' did not belong to the topics of Section 3 and is in Section 4. Also, pre-trained models are handled in an own subsection of Section 4 to make the parallel with starting from scratch visible.