setup.md: clarify usage of conda pip and OSs
Conda and pip work regardless of the operating system once their installers have been successfully.

In the previous version, the commands to run conda/pip were framed as bash commands.
This implied that the users had to use a Unix Shell proper or a simulator of it, which is not needed;
in turn this might lead to the impression that troubleshooting installation issues is harder than it is.

Therefore, the instructions to install packages have been framed as generic source code. 
Some context information clarifies the relationship between Anaconda, Python, Conda, pip and the operating system.
Other minor spelling glitches have been fixed.